[prometheus-metrics-adapter] Add logs and resync time x2
