+++ modules/301-prometheus-metrics-adapter/images/prometheus-reverse-proxy/Dockerfile
@@ -1,9 +1,10 @@
 ARG BASE_ALPINE
-ARG BASE_GOLANG_16_ALPINE
-FROM $BASE_GOLANG_16_ALPINE as artifact
+ARG BASE_GOLANG_17_ALPINE
+FROM $BASE_GOLANG_17_ALPINE as artifact
 WORKDIR /src/
-COPY main.go go.mod go.sum /src/
-RUN GOOS=linux GOARCH=amd64 CGO_ENABLED=0 go build -ldflags="-s -w" -o prometheus-reverse-proxy main.go
+COPY /app/ /src/
+RUN apk add --no-cache git && \
+    GOOS=linux GOARCH=amd64 CGO_ENABLED=0 go build -ldflags="-s -w" ./cmd/prometheus-reverse-proxy
 
 FROM $BASE_ALPINE
 COPY --from=artifact /src/prometheus-reverse-proxy /prometheus-reverse-proxy
