+++ modules/301-prometheus-metrics-adapter/images/prometheus-reverse-proxy/app/go.mod
@@ -1,0 +1,5 @@
+module prometheus-reverse-proxy
+
+go 1.17
+
+require github.com/google/uuid v1.3.0
