--- modules/301-prometheus-metrics-adapter/images/prometheus-reverse-proxy/go.mod
@@ -1,3 +1,0 @@
-module prometheus-reverse-proxy
-
-go 1.12
